refactor: mover productos a features/productos/
- Mover componentes, hooks, domain types y pagina de productos
  a src/features/productos/ siguiendo estructura feature-based
- Actualizar imports en Dashboard, routes, useCarrito
- Eliminar carpeta obsoleta src/components/products/
