+++ locuventas_frontend/src/features/productos/components/GestionProductos.tsx
@@ -1,13 +1,13 @@
 import { useState } from "react";
 import type { SelectOption } from "@domain/ui.types";
-import useProductos from "@hooks/useProductos";
-import useFiltrosProducto from "@hooks/useFiltrosProducto";
-import useGestionProductos from "@hooks/useGestionProductos";
+import useProductos from "../hooks/useProductos";
+import useFiltrosProducto from "../hooks/useFiltrosProducto";
+import useGestionProductos from "../hooks/useGestionProductos";
 import useBreakpoint from "@hooks/useBreakpoint";
 import { isBreakpoint } from "@constants/breakpoints";
-import TablaProductos from "@components/products/TablaProductos";
-import ProductoGestionCard from "@components/products/ProductoGestionCard";
-import ModalProductoForm from "@components/products/ModalProductoForm";
+import TablaProductos from "./TablaProductos";
+import ProductoGestionCard from "./ProductoGestionCard";
+import ModalProductoForm from "./ModalProductoForm";
 import ModalConfirmacion from "@components/common/ModalConfirmacion";
 import Paginacion from "@components/common/Paginacion";
 import BuscadorInput from "@components/common/BuscadorInput";
