+++ locuventas_frontend/src/features/productos/domain/producto.types.ts
